Began remote container support + tiny refactors
Added in the ability to specify the Postgres/RabbitMQ host and creds. Initially I wanted this so we can deploy without having to use docker "host" networking and to remove the hardcoded creds that are currently in the config files. WSL doesn't support docker host networking and host networking isn't conducive for deployments across multiple hosts. The changes needed for this aren't complete yet, but it's mostly there (just need to pass the environment variables on to the C2/payload containers now).

Also made some changes to the startup scripts and certain log messages to preventing printing "errors" that aren't really errors.

An outstanding thing is we need to update the RabbitMQ docker image. With the changes I made, there's no longer a need to add a user at startup to RabbitMQ. Instead, we just let the official RabbitMQ container handle that for us. So we can probably get rid of the init/config_rabbit scripts since they're no longer needed.

Changelog:
- Refactored arg parsing in start_mythic.sh
- start_mythic.sh now validate that the config.json file is properly formatted JSON
- Can now specify postgres_host, postgres_port, postgres_db, postgres_user, and postgres_password from the config.json
- start_mythic.sh generates the .env file dynamically
- Changed the .gitignore to ignore the .env and display_output.txt files
- Changed the Mythic Server's entrypoint in docker-compose.yml. wait-for-it.sh used to always start the server, even if Postgres wasn't online. It now waits 30 seconds and will not attempt to start the server if the port isn't open.
- Made ./start_documentation.sh only delete the container if it exists (was causing errors when running start_mythic.sh)
- Removed Mythic Server's HTTP support (only supports HTTPS now)
- Mythic Server waits for RabbitMQ to start gracefully rather than throwing errors in the logs when it is starting up
- config.json now has a "debug" option that starts Sanic in debug mode. Provides debug logs + auto reloading on python file changes.
- Fixed username enumeration bug due to differences in responses on the login page
- Added logging around failed authentication
- Updated several locations to use Sanic's logging rather than printing to the console
- Added the wait-for-it license to the LICENSE file
- Can now specify rabbitmq_host, rabbitmq_port, rabbitmq_user, rabbitmq_password, and rabbitmq_vhost from the config.json.
  - Note: this is note completely functional yet, but the defaults work just fine. Still need to propagate these config values to C2/payload containers.
- RabbitMQ's password is now randomized if one is not specified
- In the RabbitMQ container, replaced the default guest:guest user with the mythic user. Also updated the RabbitMQ docker template to no longer create the mythic user (not needed since we can use the RabbitMQ container's RABBITMQ_DEFAULT_USER/RABBITMQ_PASSWORD/RABBITMQ_VHOST environment variables)