Unify install_docker_X scripts
This update merges the "install_docker_$DISTRO" scripts. It basically uses lsb_release to determine the distro and wraps each of the existing scripts in their own functions.

Two changes to the scripts as well: The Ubuntu and Debian versions now download the latest version of docker-compose -- and -- the kali script no longer forces a package upgrade (apt-get upgrade -y)